feat(frontend): review icon arrow up
