--- src/frontend/src/lib/components/icons/IconArrowCircleUp.svelte
@@ -1,19 +1,0 @@
-<!-- source: https://fonts.google.com/icons?selected=Material+Symbols+Rounded:arrow_circle_up:FILL@0;wght@300;GRAD@0;opsz@24&icon.size=24&icon.color=%23e8eaed&icon.query=arrow_circle_up&icon.set=Material+Symbols&icon.style=Rounded -->
-<script lang="ts">
-	interface Props {
-		size?: string;
-	}
-
-	let { size = '48px' }: Props = $props();
-</script>
-
-<svg
-	fill="currentColor"
-	height={size}
-	viewBox="0 -960 960 960"
-	width={size}
-	xmlns="http://www.w3.org/2000/svg"
-	><path
-		d="M450-514.54v152.23q0 12.75 8.63 21.38 8.63 8.62 21.38 8.62 12.76 0 21.57-8.73 8.8-8.74 8.8-21.65v-151.85l53.31 53.31q8.31 8.31 20.89 8.11 12.57-.19 20.88-8.5 8.31-8.3 8.31-20.94t-8.31-20.82l-99.77-99.39q-10.84-10.84-25.31-10.84-14.46 0-25.3 10.84L355.31-503q-8.31 8.31-8.31 20.69 0 12.39 8.31 20.69 8.31 8.31 20.95 8.31 12.64 0 20.82-8.31L450-514.54ZM480.07-100q-78.84 0-148.21-29.92t-120.68-81.21q-51.31-51.29-81.25-120.63Q100-401.1 100-479.93q0-78.84 29.92-148.21t81.21-120.68q51.29-51.31 120.63-81.25Q401.1-860 479.93-860q78.84 0 148.21 29.92t120.68 81.21q51.31 51.29 81.25 120.63Q860-558.9 860-480.07q0 78.84-29.92 148.21t-81.21 120.68q-51.29 51.31-120.63 81.25Q558.9-100 480.07-100Zm-.07-60q134 0 227-93t93-227q0-134-93-227t-227-93q-134 0-227 93t-93 227q0 134 93 227t227 93Zm0-320Z"
-	/></svg
->
